chore: add install make target to install provider
Add a make target to install the provider so that the Terraform CLI can find it and be used for dev
purposes. It installs the plugin in $HOME/.terraform/d/plugins so that Terraform can produce an
implied configuration which includes a filesystem mirror to find the local dev builds of the
provider.